Add license file that ships with the Gem
A lot of people run license scanners in their projects to ensure they're
compliant with corporate license policies. I added a license file that
ships with the gem to help those folks out. I also fixed the readme
badges and some readme typos while I was in there. I'd greatly
appreciate a merge / release of this.

Signed-off-by: Tim Smith <tsmith@chef.io>